COMMENT: Added comments clarifying the previous WAR
It looks like the WAR  in commit 44477d33ff641663b0e6b0716c45ac2cb5fab349
is wxWidgets issue 3494 http://trac.wxwidgets.org/ticket/3494

The problem is when a DIB is selected onto a DC and a
StretchBlt sources from that DC, the units used need to be
device units but wx uses logical units.

This matches the WAR observations, where using a scratch
BMP instead of the incoming image (which is likely a DIB)
fixes the issue.